controller moved to the 'routes' folder
