+++ src/index.ts
@@ -4,8 +4,8 @@ import express = require("express")
 import * as expressPino from "express-pino-logger"
 import * as mongoose from "mongoose"
 import pino from "pino"
-import AuthenticationController from "./authentication/authentication.controller"
 import filmRouter from "./controllers/filmController/film.controller"
+import AuthenticationController from "./routes/authentication.controller"
 import errorMiddleware from "./utils/middlewares/error.middleware"
 import "dotenv/config"
 import validateEnv from "./utils/validateEnv"
